requirements.txt
affine==2.1.0
again==1.2.21
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cairocffi==0.8.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
cycler==0.10.0
decorator==4.0.11
Django==1.10.5
docopt==0.6.2
enum34==1.1.6
functools32==3.2.3.post2
GDAL==1.11.5
GeoAlchemy2==0.4.0
geopy==1.11.0
googlemaps==2.5.0
idna==2.5
ipython==5.4.1
ipython-genutils==0.2.0
jsonstreamer==1.3.6
lazyme==0.0.15
matplotlib==2.0.2
MySQL-python==1.2.5
numpy==1.13.1
olefile==0.44
pandas==0.20.3
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pipreqs==0.4.9
prompt-toolkit==1.0.14
psycopg2==2.7.3
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
pyshp==1.2.11
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
rasterio==0.36.0
requests==2.18.1
scandir==1.5
scipy==0.19.1
Shapely==1.5.17.post1
simplegeneric==0.8.1
six==1.10.0
snuggs==1.4.1
SQLAlchemy==1.1.13
subprocess32==3.2.7
svgwrite==1.1.11
traitlets==4.3.2
ujson==1.35
urllib3==1.21.1
wcwidth==0.1.7
yarg==0.1.9
